moved configureEndpoint function into separate file to be used by both fastify and express
